% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnu_CMAUPv1.R
\name{dnu_CMAUPv1}
\alias{dnu_CMAUPv1}
\title{Get Diseases, Numbers and URLs from CMAUP v1.0}
\usage{
dnu_CMAUPv1(outpath = NULL)
}
\arguments{
\item{outpath}{path for the output files. When outpath=NULL (default) the
current working directory is used.}
}
\value{
A text file ("Diseases_Numbers_URLs_v1") with three columns named
"Disease name", "Disease_number" and "URL", respectively. Values are
separate by comma and space.
}
\description{
Scraping disease names, disease numbers and disease URLs from
CMAUP v1.0.
}
\details{
This function retrieve names and URLs of all the diseases included
on the CMAUP v1.0 website. Since diseases are archived following an
internal numbering (please note that disease names and  numbering are
different between CMAUP v1.0 and CMAUP v2.0),the number corresponding to
each disease is also provided. These numbers are useful to set the "dn"
parameter in the 'pr2d_CMAUPv1' function.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Use tempdir() to find the current temp directory.
library(rvest)
library(polite)
# Use a temporary directory.
path<-paste0(normalizePath(tempdir(), winslash = "/"),"/")
# Scraping data from CMAUP v1.0
dnu_CMAUPv1(outpath=path)
#Remove the output file from the current temporary directory.
toremove<-"Diseases_Numbers_URLs_v1"
file.remove(paste0(path,toremove))

\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item Perepolkin D. (2023) polite: Be Nice on the Web.
https://github.com/dmi3kno/polite \cr
\item Wickham H (2023). rvest: Easily Harvest (Scrape) Web Pages.
https://rvest.tidyverse.org/, https://github.com/tidyverse/rvest. \cr
\item Xian Zeng, Peng Zhang, Yali Wang, et al. CMAUP: a database of collective
molecular activities of useful plants. Nucleic Acids Research 2019;
47(D1): D1118-D1127  DOI:doi.org/10.1093/nar/gky965.
\item Zecca, G., Toini, E., Labra, M, Grassi, F. (2025) Accelerating the
identification and the prioritisation of new plants with medicinal
potential: the  pm4mp R package.\cr
}
}
\seealso{
\code{\link[=pr2d_CMAUPv1]{pr2d_CMAUPv1()}}, \code{\link[=dnu_CMAUPv2]{dnu_CMAUPv2()}}
}
\author{
Giovanni Zecca, \email{giovanni.zecca@unimib.it}
}
