% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2d_CMAUPv1.R
\name{pr2d_CMAUPv1}
\alias{pr2d_CMAUPv1}
\title{Scraping Plants Related to Diseases (CMAUP v1.0)}
\usage{
pr2d_CMAUPv1(ALL = TRUE, MED = TRUE, dn = NULL, outpath = NULL)
}
\arguments{
\item{ALL}{Logical (default=TRUE), indicating whether files including all
plants related to the required diseases must be created.}

\item{MED}{Logical (default=TRUE), indicating whether files including only
the "Medicinal Plants" (as defined by the CMAUP's Plant Usage Classes) must
be created for the requested diseases.}

\item{dn}{Integer vector, specifying the diseases of interest. Diseases are
selected using the internal numbering used by CMAUP v1.0. The number
corresponding to each disease can be easily obtained using the
'dnu_CMAUPv1' function. Please note that disease names and numbers differ
between CMAUP v1.0 and CMAUP v2.0. When dn= NULL (default), information
relating to all the diseases present on the CMAUP v.10 website are included
in downloads, according to the values of the ALL and MED parameters. Keep
in mind that in this case the function may take a long time to complete its
work (polite webscraping is applied and hundreds of file are created). See
examples below.}

\item{outpath}{path for the output files. When outpath=NULL (default) the
current working directory is used.}
}
\value{
For each disease a three-columns csv file is created containing
information relating to the name, genus and family of plants linked to that
disease. Values are separate by comma. Files are named according to the
following scheme: "Disease name_ICD-10 code_ALL/MED_number of plants
included in the file_CMAUP1.csv"
}
\description{
The lists of plants associated with the requested diseases are
retrieved from CMAUP v1.0 site.
}
\details{
This function creates several files storing lists of plants
associated with the required diseases. Please note that at least one of the
parameters ALL and MED must be set as TRUE.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(rvest)
library(polite)

 # NOT RUN!
 # All diseases present on CMAUP v1.0 are selected. If n is the number of
 # diseases, 2n file are created (i.e., n_ALL + n_MED). It may take a long time
 # to complete!

 # pr2d_CMAUPv1()

 # All diseases (n) present on CMAUP v1.0 are selected, but only the n_MED
 # files are created. It may take a long time to complete!

 # pr2d_CMAUPv1(ALL=FALSE)

 # All diseases (n) present on CMAUP v1.0 are selected, but only the n_ALL
 # files are created. It may take a long time to complete!

 # pr2d_CMAUPv1(MED=FALSE)

 # RUN
 # Use tempdir() to find the current temp directory.

 path<-paste0(normalizePath(tempdir(), winslash = "/"),"/")

 # Only disease#1 is selected and only one file is created in the temporary folder.

 pr2d_CMAUPv1(dn=c(1), MED=FALSE, outpath=path)

 # Only disease# 2,7 and 25 are selected and only 3_ALL files are
 # created in the temporary folder.

 pr2d_CMAUPv1(dn=c(2,7,25), MED=FALSE, outpath=path)

 #Remove the output file from the current temporary directory.
 toremove<-list.files(path, pattern = "*_CMAUP1.csv$")
 file.remove(paste0(path,toremove))
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item Perepolkin D. (2023) polite: Be Nice on the Web.
https://github.com/dmi3kno/polite \cr
\item Wickham H (2023). rvest: Easily Harvest (Scrape) Web Pages.
https://rvest.tidyverse.org/, https://github.com/tidyverse/rvest. \cr
\item Xian Zeng, Peng Zhang, Yali Wang, et al. CMAUP: a database of collective
molecular activities of useful plants. Nucleic Acids Research 2019;
47(D1): D1118-D1127  DOI: doi.org/10.1093/nar/gky965.
\item Zecca, G., Toini, E., Labra, M, Grassi, F. (2025) Accelerating the
identification and the prioritisation of new plants with medicinal
potential: the  pm4mp R package.\cr
}
}
\seealso{
\link{dnu_CMAUPv1}
}
\author{
Giovanni Zecca, \email{giovanni.zecca@unimib.it}
}
