% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree4nodesigl.R
\name{tree4nodesigl}
\alias{tree4nodesigl}
\title{Preparing the input 'tree file' for Phylocom}
\usage{
tree4nodesigl(tree, fileout)
}
\arguments{
\item{tree}{A user-provided tree in Newick format whose tips are assumed to
be correctly labelled. See 'clean_and_match' function for details and
options.}

\item{fileout}{Character string specifying the suffix to add to the output
files. This can be the same name as the input tree file or any other name
chosen by the user.}
}
\value{
Two tree files are saved in Newick format, which include modified
version of the input tree. The first one, named
"phylo_nobl_fileout",contains a tree without branch lengths and with
internal nodes named as explained in Details section. The latter, named
"phylo_bl_fileout", is the same as the former but it includes branch
lengths.
These trees are intended for use in Phylocom's"nodesigl"analysis (no branch lengths)
and subsequent "has" and "hmpp" computations (with branch lengths).
}
\description{
This function prepares the input tree to be used with the
software Phylocom (Webb et al. 2008) starting from a tree provided by the
user.
}
\details{
The functions assumes that the user-provided tree has correctly
labelled tips. See 'clean_and_match' function for details and renaming
options. Unnamed internal nodes (i.e, NULL, "", "NA" or NA) are named
following the numbering scheme used by ape package (i.e., from Ntips+1 to
Ntips+Nnode), with numbers preceded by the prefix "ApeNumb.". Existing
names are however preserved.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Use tempdir() to find the current temp directory.
 library(ape)
# Generate a random tree with 10 tips.
tr<-rtree(10)

# Name internal nodes 3,4 and 9 with labels "A","B","C". Other nodes are
# named with "" or NA.
  tr$node.label[c(3,4,9)]<- LETTERS[1:3]
  tr$node.label[1:2]<-""

 # See internal node labels
  tr$node.label

 #Write the tree in the current temp folder.
  tree<-write.tree(tr)
  treefile<- tempfile("tree", fileext = ".tree")
  cat(tree, file = treefile, sep ="\n")

 # Get path to the current temp folder and set it as the current working
 # directory.
  path<-paste0(normalizePath(tempdir(), winslash = "/"),"/")
  setwd(path)

 #Test.
  tree4nodesigl(treefile, "random_tree")

 #Tidy up. Remove the output files from the current temporary directory.
  toremove<-list.files(path, pattern = ".tree|phylo_")
  file.remove(paste0(path,toremove))
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item Paradis E, Schliep K (2019). “ape 5.0: an environment for modern phylogenetics
and evolutionary analyses in R.”
Bioinformatics, 35, 526-528. doi:10.1093/bioinformatics/bty633.\cr
\item Webb, C. O.; Ackerly, D. D. & Kembel, S. W. (2008) Phylocom: software for the
analysis of phylogenetic community structure and trait evolution.
Bioinformatics 24: 2098-2100.
\item Zecca, G., Toini, E., Labra, M, Grassi, F. (2025) Accelerating the
identification and the prioritisation of new plants with medicinal
potential: the  pm4mp R package.\cr
}
}
\seealso{
\code{\link[=clean_and_match]{clean_and_match()}}, \code{\link[=sample4nodesigl]{sample4nodesigl()}}
}
\author{
Giovanni Zecca, \email{giovanni.zecca@unimib.it}
}
